+++ input/rabbitmq/rabbitmq.go
@@ -19,10 +19,7 @@ const MessageIdFormat = "senzing-6201%04d"
 
 // ----------------------------------------------------------------------------
 // TODO: rename?
-func Read(urlString string, delay time.Duration) {
-
-	fmt.Println(time.Now(), "Sleep for 60 seconds to let RabbitMQ and Postgres settle down and come up.")
-	time.Sleep(delay * time.Minute)
+func Read(urlString string) {
 
 	ctx := context.Background()