+++ input/selector.go
@@ -4,7 +4,6 @@ import (
 	"fmt"
 	"net"
 	"net/url"
-	"time"
 
 	"github.com/roncewind/load/input/rabbitmq"
 	"github.com/senzing/go-logging/messagelogger"
@@ -69,7 +68,7 @@ func parseURL(urlString string) *url.URL {
 }
 
 // ----------------------------------------------------------------------------
-func Read(inputURL, logLevel string, delay time.Duration) bool {
+func Read(inputURL, logLevel string) bool {
 	if len(logLevel) > 0 {
 		msglog.SetLogLevelFromString(logLevel)
 	}
@@ -78,7 +77,7 @@ func Read(inputURL, logLevel string, delay time.Duration) bool {
 	switch u.Scheme {
 	case "amqp":
 		if len(inputURL) > 0 {
-			rabbitmq.Read(inputURL, delay)
+			rabbitmq.Read(inputURL)
 		} else {
 			return false
 		}